refacto: uses tm2/commands instead of stdlib flags
