+++ contribs/github-bot/go.mod
@@ -1,8 +1,11 @@
 module github.com/gnolang/gno/contribs/github-bot
 
-go 1.21
+go 1.22
+
+toolchain go1.22.2
 
 require (
+	github.com/gnolang/gno v0.2.0
 	github.com/google/go-github/v64 v64.0.0
 	github.com/migueleliasweb/go-github-mock v1.0.1
 	github.com/sethvargo/go-githubactions v1.3.0
@@ -11,6 +14,9 @@ require (
 
 require (
 	github.com/google/go-querystring v1.1.0 // indirect
-	github.com/gorilla/mux v1.8.0 // indirect
+	github.com/gorilla/mux v1.8.1 // indirect
+	github.com/peterbourgon/ff/v3 v3.4.0 // indirect
+	golang.org/x/sys v0.22.0 // indirect
+	golang.org/x/term v0.22.0 // indirect
 	golang.org/x/time v0.3.0 // indirect
 )
